refactor: move buffer abstraction into its own crate
