+++ src/protocols/whir_zk/prover.rs
@@ -1,3 +1,5 @@
+// IGNORE CHANGES TO THIS FILE - NOT FULLY PORTED TO PROPERLY USE BUFFER ABSTRACTION.
+
 use std::borrow::Cow;
 
 use ark_ff::Field;
@@ -8,8 +10,10 @@ use rayon::prelude::*;
 use tracing::instrument;
 
 use super::{Config, Witness};
+use crate::buffer::BufferOps;
 use crate::{
     algebra::{
+        buffer::ActiveBuffer,
         embedding::Identity,
         linear_form::{Covector, Evaluate, LinearForm},
         mixed_dot, scalar_mul_add,
@@ -209,7 +213,7 @@ impl<F: Field> Config<F> {
     pub fn prove<'a, H, R>(
         &self,
         prover_state: &mut ProverState<H, R>,
-        vectors: Vec<Cow<'a, [F]>>,
+        vectors: &[&ActiveBuffer<F>],
         witness: Witness<F>,
         linear_forms: Vec<Box<dyn LinearForm<F>>>,
         evaluations: Cow<'a, [F]>,
@@ -265,7 +269,6 @@ impl<F: Field> Config<F> {
         let num_witness_variables = self.num_witness_variables();
         let num_blinding_variables = self.num_blinding_variables();
         let num_witness_variables_plus_1 = num_witness_variables + 1;
-        drop(vectors); // TODO: These are never touched?
 
         // Compute w_folded evaluations of all blinding vectors before rho for binding.
         let (w_folded_weights, m_evals, w_folded_blinding_evals) = {
@@ -396,10 +399,16 @@ impl<F: Field> Config<F> {
         let result = {
             #[cfg(feature = "tracing")]
             let _span = tracing::info_span!("inner_blinded_prove").entered();
+            let f_hat_buffers = f_hat_vectors
+                .iter()
+                .map(|v| ActiveBuffer::from_slice(v))
+                .collect::<Vec<_>>();
+            let f_hat_refs = f_hat_buffers.iter().collect::<Vec<_>>();
+            let f_hat_witness_refs = f_hat_witnesses.iter().collect::<Vec<_>>();
             self.blinded_commitment.prove(
                 prover_state,
-                f_hat_vectors.into_iter().map(Cow::Owned).collect(),
-                f_hat_witnesses.into_iter().map(Cow::Owned).collect(),
+                &f_hat_refs,
+                f_hat_witness_refs,
                 linear_forms,
                 Cow::Owned(modified_evaluations),
             )
@@ -423,10 +432,15 @@ impl<F: Field> Config<F> {
             .concat();
             // Blinding sub-proof result is discarded: the blinding WHIR's
             // evaluation point is not needed by the outer protocol.
+            let blinding_buffers = blinding_vectors
+                .iter()
+                .map(|v| ActiveBuffer::from_slice(v))
+                .collect::<Vec<_>>();
+            let blinding_refs = blinding_buffers.iter().collect::<Vec<_>>();
             let _ = self.blinding_commitment.prove(
                 prover_state,
-                blinding_vectors.into_iter().map(Cow::Owned).collect(),
-                vec![Cow::Owned(blinding_witness)],
+                &blinding_refs,
+                vec![&blinding_witness],
                 blinding_forms,
                 Cow::Owned(all_blinding_claims),
             );
